// web/src/lib/types.ts
export interface StandardAnswerCategory {
  id: number;
  name: string;
}

export interface StandardAnswer {
  id: number;
  keyword: string;
  answer: string;
  categories: StandardAnswerCategory[];
}